requirements.txt
# Core dependencies
google-api-python-client==2.100.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.0.0
google-auth==2.23.0

# PDF processing
PyPDF2==3.0.1
pdfplumber==0.9.0
pymupdf==1.23.5

# OCR capabilities
pytesseract==0.3.10
Pillow==10.0.0
pdf2image==1.16.3

# Hindi text processing and transliteration
indic-transliteration==2.3.42
regex==2023.8.8

# Data processing
pandas==2.1.1
numpy==1.24.3

# Utilities
python-dotenv==1.0.0
schedule==1.2.0
requests==2.31.0
pathlib2==2.3.7

# Testing
pytest==7.4.2
pytest-cov==4.1.0
pytest-mock==3.11.1

# Logging
colorlog==6.7.0
